video: tegra: nvmap: change ivmid to 64bit
- the id field in nvmap_create_handle at present is of 32bit,
which is also used for ivm ioctl's,this change is required as
we need to increase the ivmid to 64bit.
- increasing the id field would increase the size of structure
and would break the compatibility of older userspace with newer
kernel.
- so, to maintain the compatibility fields in nvmap_create_handle
structure have been adjusted such that overall size still remain same.

bug 1672120

Reviewed-on: http://git-master/r/1218353

+++ drivers/video/tegra/nvmap/nvmap_handle.c
@@ -154,7 +154,7 @@ ref_alloc_fail:
 }
 
 struct nvmap_handle_ref *nvmap_try_duplicate_by_ivmid(
-		struct nvmap_client *client, unsigned int ivm_id,
+		struct nvmap_client *client, u64 ivm_id,
 		struct nvmap_heap_block **block)
 {
 	struct nvmap_handle *h = NULL;
